Update longest_substring_without_repeating_characters.py
Improved solve - this solution uses a window approach instead of brute force through each position in the string.  Instead, when we find a repeat character, we slice and keep the substring that we know is unique, removing the prev repeat character, and appending the new repeated character and continuing one.